requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
google-generativeai>=0.7.2
google-api-python-client>=2.130.0
